requirements.txt
Django
django-rest-knox
djangorestframework
python-decouple
raven
django-cors-headers